splitted some files to better organization
