+++ packages/app/data/schema.graphql
@@ -26,6 +26,9 @@ type Mutation {
   EventCreate(input: EventCreateInput!): EventCreatePayload
 }
 
+"""
+The root of all... queries
+"""
 type Query {
   event(id: ID!): Event
   events: [Event]
