+++ packages/server/src/modules/event/mutation/index.js
@@ -1,0 +1,5 @@
+import EventCreate from "./EventCreateMutation";
+
+export default {
+  EventCreate
+};
