+++ packages/server/src/schema.js
@@ -1,0 +1,9 @@
+import { GraphQLSchema } from "graphql";
+
+import QueryType from "./type/QueryType";
+import MutationType from "./type/MutationType";
+
+export const schema = new GraphQLSchema({
+  query: QueryType,
+  mutation: MutationType
+});
