+++ packages/server/src/server.js
@@ -1,7 +1,7 @@
 const Koa = require("koa");
 const mount = require("koa-mount");
 const graphqlHTTP = require("koa-graphql");
-const schema = require("./graphql/schema");
+const schema = require("./schema");
 const mongoose = require("mongoose");
 
 const databaseUrl = "mongodb://127.0.0.1:27017/test";
