+++ packages/server/src/type/MutationType.js
@@ -1,0 +1,10 @@
+import { GraphQLObjectType } from "graphql";
+
+import EventMutations from "../modules/event/mutation";
+
+export default new GraphQLObjectType({
+  name: "Mutation",
+  fields: () => ({
+    ...EventMutations
+  })
+});
